# web/static/openapi.yaml
openapi: "3.1.0"
info:
  title: lima-container API
  version: "1.0"
  description: |
    REST API exposed by the `lima-web` Go server inside lima-container images.
    Available in the `lima-web` and `lima-bootc` image variants.
    All responses use `{"data": ...}` on success or `{"error": "..."}` on failure.
  license:
    name: MIT
servers:
  - url: http://localhost:8006
    description: Local container
...
paths:
...
  /api/instances/{name}/rdp:
    get:
      operationId: getRDPStatus
      summary: Get RDP status for a VM instance
      description: Check if RDP (GNOME Remote Desktop or xrdp) is available and enabled.
      tags: [rdp]
      parameters:
        - $ref: "#/components/parameters/instanceName"
      responses:
        "200":
          description: RDP status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RDPStatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  parameters:
    instanceName:
      name: name
      in: path
      required: true
      schema:
        type: string
      description: Lima instance name
...
  responses:
...
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
...
    RDPStatus:
      type: object
      required: [available, type, port, enabled]
      properties:
        available:
          type: boolean
          description: Whether an RDP server is installed in the VM
        type:
          type: string
          enum: [grd, xrdp, none]
          description: Type of RDP server detected
        port:
          type: integer
          description: RDP port (typically 3389)
          example: 3389
        enabled:
          type: boolean
          description: Whether port 3389 is currently listening
    RDPStatusResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/RDPStatus"
...